test: ProgressReview 구현체들 테스트 코드 작성
