+++ backend/src/test/java/com/wootech/dropthecode/domain/review/CompletedReviewTest.java
@@ -1,0 +1,62 @@
+package com.wootech.dropthecode.domain.review;
+
+import com.wootech.dropthecode.domain.*;
+import com.wootech.dropthecode.exception.AuthorizationException;
+import com.wootech.dropthecode.exception.ReviewException;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+import static com.wootech.dropthecode.builder.MemberBuilder.dummyMember;
+import static com.wootech.dropthecode.builder.ReviewBuilder.dummyReview;
+import static org.assertj.core.api.Assertions.assertThatCode;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+
+public class CompletedReviewTest {
+
+    @DisplayName("Finished 상태 변경이 가능한 경우 - 성공")
+    @Test
+    void updateFinishedReview() {
+        Member teacher = dummyMember(1L, "1000", "Jinho", "jh8579@gmail.com", "s3://jh8579", "github url", Role.TEACHER, null);
+        Member student = dummyMember(2L, "1000", "Jinho", "jh8579@gmail.com", "s3://jh8579", "github url", Role.STUDENT, null);
+
+        Review review = dummyReview(teacher, student, "test title", "test content", "github/3", 0L, Progress.TEACHER_COMPLETED);
+
+        assertThatCode(() -> new CompletedReview(review).finish(2L)).doesNotThrowAnyException();
+    }
+
+    @DisplayName("Finished 상태 변경이 가능한 경우 - 실패")
+    @Test
+    void updateFinishedReviewFail() {
+        Member teacher = dummyMember(1L, "1000", "Jinho", "jh8579@gmail.com", "s3://jh8579", "github url", Role.TEACHER, null);
+        Member student = dummyMember(2L, "1000", "Jinho", "jh8579@gmail.com", "s3://jh8579", "github url", Role.STUDENT, null);
+
+        Review review = dummyReview(teacher, student, "test title", "test content", "github/3", 0L, Progress.ON_GOING);
+
+        assertThatCode(() -> new CompletedReview(review).finish(2L)).isInstanceOf(ReviewException.class);
+    }
+
+    @DisplayName("Finished 상태 변경이 가능한 경우 - 학생 ID 실패")
+    @Test
+    void validateStudentIdFail() {
+        Member teacher = dummyMember(1L, "1000", "Jinho", "jh8579@gmail.com", "s3://jh8579", "github url", Role.TEACHER, null);
+        Member student = dummyMember(2L, "1000", "Jinho", "jh8579@gmail.com", "s3://jh8579", "github url", Role.STUDENT, null);
+
+        Review review = dummyReview(teacher, student, "test title", "test content", "github/3", 0L, Progress.TEACHER_COMPLETED);
+
+        assertThatThrownBy(() -> review.validateAuthorityOfStudent(teacher.getId()))
+                .isInstanceOf(AuthorizationException.class);
+    }
+
+    @DisplayName("Finished 상태 변경이 가능한 경우 - 선생님 ID 실패")
+    @Test
+    void validateTeacherIdFail() {
+        Member teacher = dummyMember(1L, "1000", "Jinho", "jh8579@gmail.com", "s3://jh8579", "github url", Role.TEACHER, null);
+        Member student = dummyMember(2L, "1000", "Jinho", "jh8579@gmail.com", "s3://jh8579", "github url", Role.STUDENT, null);
+
+        Review review = dummyReview(teacher, student, "test title", "test content", "github/3", 0L, Progress.TEACHER_COMPLETED);
+
+        assertThatThrownBy(() -> review.validateAuthorityOfTeacher(student.getId()))
+                .isInstanceOf(AuthorizationException.class);
+    }
+}
